Reorganize price estimation into module
