+++ shared/src/arguments.rs
@@ -1,6 +1,6 @@
 //! Contains command line arguments and related helpers that are shared between the binaries.
 use crate::{
-    gas_price_estimation::GasEstimatorType, price_estimate::PriceEstimatorType,
+    gas_price_estimation::GasEstimatorType, price_estimation::PriceEstimatorType,
     sources::BaselineSource,
 };
 use anyhow::{ensure, Result};
