+++ shared/src/lib.rs
@@ -13,8 +13,7 @@ pub mod maintenance;
 pub mod metrics;
 pub mod network;
 pub mod paraswap_api;
-pub mod paraswap_price_estimator;
-pub mod price_estimate;
+pub mod price_estimation;
 pub mod recent_block_cache;
 pub mod sources;
 pub mod subgraph;
